new matrix editor replaces old one
